endian order with type template
