+++ tkcore/core/CMakeLists.txt
@@ -1,12 +1,10 @@
 
-
-
-
 add_library(key_core
         # Core
         ${CMAKE_CURRENT_LIST_DIR}/common.cpp
         ${CMAKE_CURRENT_LIST_DIR}/key_signatures.cpp
         ${CMAKE_CURRENT_LIST_DIR}/key_errors.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/key_endian.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt/salt_base.cpp
 )
