+++ tkcore/core/key_endian.h
@@ -10,23 +10,15 @@ enum endian {
     bigEndian
 };
 
-endian currentEndian() {
-    int num = 1;
-    if (*(char *) &num == 1) {
-        return littleEndian;
-    } else {
-        return bigEndian;
-    }
-
-}
+endian currentEndian();
 
 template<typename T>
 T swap(const T &value, const endian &from, const endian &to) {
     if (from == to) {
         return value;
     } else {
-        auto *origin = (const uint8_t *) &value;
-        uint8_t byte[sizeof(T)];
+        auto *origin = (const char *) &value;
+        char byte[sizeof(T)];
         for (int i = 0; i < sizeof(value); ++i) {
             byte[sizeof(T) - i - 1] = origin[i];
         }
@@ -37,6 +29,17 @@ T swap(const T &value, const endian &from, const endian &to) {
 template<typename T>
 T swap(const T &origin, const endian &to) {
     return swap(origin, currentEndian(), to);
+}
+
+/**
+ * Host to big endian
+ * @tparam T
+ * @param origin
+ * @return
+ */
+template<typename T>
+T htobig(const T &origin) {
+    return swap(origin, currentEndian(), bigEndian);
 }
 
 #endif //THEKEY_ENDIAN_H
