+++ tkcore/core/key_signatures.h
@@ -24,18 +24,18 @@
 
 #define INT32_BIG_ENDIAN(name) \
     uint32_t raw##name;        \
-    [[nodiscard]] uint32_t name() const { return htonl(raw##name); } \
-    void name(uint32_t name) { raw##name = htonl(name); }
+    [[nodiscard]] uint32_t name() const { return htobig( raw##name ); } \
+    void name(uint32_t name) { raw##name = htobig( name ); }
 
 #define INT64_BIG_ENDIAN(name) \
     uint64_t raw##name;        \
-    [[nodiscard]] uint64_t name() const { return htole64(raw##name); } \
-    void name(uint64_t name) { raw##name = htole64(name); }
+    [[nodiscard]] uint64_t name() const { return htobig( raw##name ); } \
+    void name(uint64_t name) { raw##name = htobig( name ); }
 
 #define INT32_BIG_ENDIAN_ENUM(name, enum) \
     uint32_t raw##name;        \
-    [[nodiscard]] enum name() const { return enum( htonl(raw##name) ); } \
-    void name(enum name) { raw##name = htonl( uint32_t(name) ); }
+    [[nodiscard]] enum name() const { return enum( htobig( raw##name ) ); } \
+    void name(enum name) { raw##name = htobig( uint32_t( name ) ); }
 
 
 namespace thekey {
